SFT-367 Removed duplicated file
